cheri: Fix compile/warnings in drivers/media/
- Skip CONFIG_DVB_VIDTV and CONFIG_VIDEO_MEDIATEK_*. These have
  embedded pointers in structs apparently used for mailbox
  communication.
- Otherwise only standard uintptr_t vs. unsigned long issues.

+++ drivers/media/pci/ivtv/ivtv-driver.h
@@ -283,7 +283,7 @@ struct ivtv_user_dma {
 };
 
 struct ivtv_dma_page_info {
-	unsigned long uaddr;
+	uintptr_t uaddr;
 	unsigned long first;
 	unsigned long last;
 	unsigned int offset;
