+++ drivers/media/pci/ivtv/ivtv-udma.h
@@ -10,7 +10,7 @@
 #define IVTV_UDMA_H
 
 /* User DMA functions */
-void ivtv_udma_get_page_info(struct ivtv_dma_page_info *dma_page, unsigned long first, unsigned long size);
+void ivtv_udma_get_page_info(struct ivtv_dma_page_info *dma_page, uintptr_t first, unsigned long size);
 int ivtv_udma_fill_sg_list(struct ivtv_user_dma *dma, struct ivtv_dma_page_info *dma_page, int map_offset);
 void ivtv_udma_fill_sg_array(struct ivtv_user_dma *dma, u32 buffer_offset, u32 buffer_offset_2, u32 split);
 int ivtv_udma_setup(struct ivtv *itv, unsigned long ivtv_dest_addr,
